+++ src/image_complete/bmp.py
@@ -1,27 +1,49 @@
-import os
 import struct
 
+from io import BytesIO
+from .base import load
 
-def is_bmp_complete(img_path):
+
+def is_bmp(img):
+    """
+    Checks whether the image represents a bitmap.
+
+    https://en.wikipedia.org/wiki/BMP_file_format#File_structure
+
+    :param img: the absolute path to the BMP image or a bytes/BytesIO object
+    :type img: str or bytes or BytesIO
+    :return: True if a bitmap
+    :rtype: bool
+    """
+    data, _ = load(img)
+    try:
+        data.seek(0)
+        header = data.read(2)
+        return header in [b"BM", b"BA", b"CI", b"CP", b"IC", b"PT"]
+    except:
+        return False
+
+
+def is_bmp_complete(img):
     """
     Checks whether the BMP image is complete.
 
     https://en.wikipedia.org/wiki/BMP_file_format#File_structure
 
-    :param img_path: the absolute path to the BMP image
-    :type img_path: str
+    :param img: the absolute path to the BMP image or a bytes/BytesIO object
+    :type img: str or bytes or BytesIO
     :return: True if complete
     :rtype: bool
     """
-
     try:
-        flen = os.path.getsize(img_path)
-        if flen > 6:
-            with open(img_path, "rb") as f:
-                f.seek(2, 0)
-                data = f.read(4)
-                blen = struct.unpack('I', data)
-                return blen[0] == flen
+        data, data_len = load(img)
+        if data is None:
+            return False
+        if data_len > 6:
+            data.seek(2, 0)
+            data = data.read(4)
+            blen = struct.unpack('I', data)
+            return blen[0] == data_len
         else:
             return False
     except: