+++ src/image_complete/gif.py
@@ -1,25 +1,46 @@
-import os
+from io import BytesIO
+from .base import load
 
 
-def is_gif_complete(img_path):
+def is_gif(img):
+    """
+    Checks whether the image represents a GIF.
+
+    https://en.wikipedia.org/wiki/GIF#File_format
+
+    :param img: the absolute path to the GIF image or a bytes/BytesIO object
+    :type img: str or bytes or BytesIO
+    :return: True if a GIF
+    :rtype: bool
+    """
+    data, _ = load(img)
+    try:
+        data.seek(0)
+        header = data.read(6)
+        return header == b"GIF89a"
+    except:
+        return False
+
+
+def is_gif_complete(img):
     """
     Checks whether the GIF image is complete.
 
     https://en.wikipedia.org/wiki/GIF#File_format
 
-    :param img_path: the absolute path to the GIF image
-    :type img_path: str
+    :param img: the absolute path to the GIF image or a bytes/BytesIO object
+    :type img: str or bytes or BytesIO
     :return: True if complete
     :rtype: bool
     """
-
     try:
-        flen = os.path.getsize(img_path)
-        if flen > 1:
-            with open(img_path, "rb") as f:
-                f.seek(flen - 1, 0)
-                marker = f.read(1)
-                return marker[0] == 59
+        data, data_len = load(img)
+        if data is None:
+            return False
+        if data_len > 1:
+            data.seek(data_len - 1, 0)
+            marker = data.read(1)
+            return marker[0] == 59
         else:
             return False
     except: