+++ tests/image_complete_tests/pngtest.py
@@ -1,6 +1,6 @@
 import unittest
 from .basetest import ImageCompleteTest
-from image_complete.png import is_png_complete
+from image_complete.png import is_png_complete, is_png
 
 
 class TestPng(ImageCompleteTest):
@@ -13,6 +13,24 @@ class TestPng(ImageCompleteTest):
 
     def test_empty_png(self):
         self.assertFalse(is_png_complete(self.data_file("empty.png")))
+
+    def test_complete_png_bytes(self):
+        self.assertTrue(is_png_complete(self.data_content("complete.png")))
+
+    def test_incomplete_png_bytes(self):
+        self.assertFalse(is_png_complete(self.data_content("incomplete.png")))
+
+    def test_empty_png_bytes(self):
+        self.assertFalse(is_png_complete(self.data_content("empty.png")))
+
+    def test_complete_is_png_bytes(self):
+        self.assertTrue(is_png(self.data_content("complete.png")))
+
+    def test_incomplete_is_png_bytes(self):
+        self.assertTrue(is_png(self.data_content("incomplete.png")))
+
+    def test_empty_is_png_bytes(self):
+        self.assertFalse(is_png(self.data_content("empty.png")))
 
 
 def suite():