+++ tests/image_complete_tests/webptest.py
@@ -1,6 +1,6 @@
 import unittest
 from .basetest import ImageCompleteTest
-from image_complete.webp import is_webp_complete
+from image_complete.webp import is_webp_complete, is_webp
 
 
 class TestPng(ImageCompleteTest):
@@ -13,6 +13,24 @@ class TestPng(ImageCompleteTest):
 
     def test_empty_webp(self):
         self.assertFalse(is_webp_complete(self.data_file("empty.webp")))
+
+    def test_complete_webp_bytes(self):
+        self.assertTrue(is_webp_complete(self.data_content("complete.webp")))
+
+    def test_incomplete_webp_bytes(self):
+        self.assertFalse(is_webp_complete(self.data_content("incomplete.webp")))
+
+    def test_empty_webp_bytes(self):
+        self.assertFalse(is_webp_complete(self.data_content("empty.webp")))
+
+    def test_complete_is_webp_bytes(self):
+        self.assertTrue(is_webp(self.data_content("complete.webp")))
+
+    def test_incomplete_is_webp_bytes(self):
+        self.assertTrue(is_webp(self.data_content("incomplete.webp")))
+
+    def test_empty_is_webp_bytes(self):
+        self.assertFalse(is_webp(self.data_content("empty.webp")))
 
 
 def suite():