+++ CHANGES.rst
@@ -1,6 +1,13 @@
 Changelog
 =========
 
+0.0.4 (2023-03-30)
+------------------
+
+- added support for determining completeness of bytes or BytesIO objects rather than just file names
+- added `is_XYZ` methods that determine whether a file/bytes/BytesIO represents file type `XYZ`
+
+
 0.0.3 (2023-03-28)
 ------------------